Fixed Player class

+++ src/demo/Player.hpp
@@ -7,16 +7,16 @@ class Player : public EventListener {
 public:
 	Player() : EventListener("Player") {}
 
-	bool OnMove(float x, float y) override { return false; }
-	bool OnPress(float x, float y) override {
+	bool OnMove(f32 x, f32 y) override { return false; }
+	bool OnPress(f32 x, f32 y) override {
 		HT_WARN("Mouse pressed at %.2f , %.2f", x, y);
 		return true;
 	}
-	bool OnRelease(float x, float y) override { return false; }
+	bool OnRelease(f32 x, f32 y) override { return false; }
 
-	bool OnScroll(float xOffset) override { return false; }
+	bool OnScroll(f32 xOffset) override { return false; }
 
-	bool OnKeyPress(int key, unsigned short modifier) override { return false; }
-	bool OnKeyRepeat(int key, unsigned short modifier) override { return false; }
-	bool OnKeyRelease(int key) override { return false; }
+	bool OnKeyPress(u16 key, u16 modifier) override { return false; }
+	bool OnKeyRepeat(u16 key, u16 modifier) override { return false; }
+	bool OnKeyRelease(u16 key) override { return false; }
 };
